# API/src/main/resources/static/swagger.yml
openapi: 3.0.0
info:
  title: PIR RDF plugin project API
  description: API documentation for the PIR RDF plugin project.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        selectedThesauri:
          type: array
          items:
            type: string
        date:
          type: string
          format: date-time
        creator:
          type: object
          properties:
            email:
              type: string
        participants:
          type: array
          items:
            type: object
            properties:
              email:
                type: string
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
        annotations:
          type: array
          items:
            $ref: "#/components/schemas/Annotation"
    Invitation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          type: object
          properties:
            email:
              type: string
    Annotation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        annotationType:
          type: string
        pageUrl:
          type: string
        turtlePath:
          type: string
        item:
          $ref: "#/components/schemas/Item"
        concepts:
          type: array
          items:
            $ref: "#/components/schemas/Concept"
        campaign:
          type: object
          properties:
            name:
              type: string
        creator:
          type: object
          properties:
            email:
              type: string
    Item:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
    Concept:
      type: object
      properties:
        value:
          type: string
        uri:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/campaign:
    post:
      tags:
        - RDF
      summary: Create a new campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                selectedThesauri:
                  type: array
                  items:
                    type: string
              required:
                - name
                - selectedThesauri
      responses:
        "201":
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "409":
          description: Conflict - duplicate campaign name
          content:
            text/plain:
              schema:
                type: string
                example: Duplicate key name. | Data integrity violation.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
    get:
      tags:
        - RDF
      summary: Get campaign by ID
      operationId: getCampaignById
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Campaign retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "403":
          description: Forbidden - user not authorized
          content:
            text/plain:
              schema:
                type: string
                example: UserNotAuthorized
        "404":
          description: Campaign not found
          content:
            text/plain:
              schema:
                type: string
                example: CampaignNotFound
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/campaign/all:
    get:
      tags:
        - RDF
      summary: Get all campaigns for current user
      operationId: getAllCampaignsByUser
      responses:
        "200":
          description: Campaigns retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
        "404":
          description: No campaigns found
          content:
            text/plain:
              schema:
                type: string
                example: CampaignsNotFound
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/campaign/invite:
    post:
      tags:
        - RDF
      summary: Invite a participant to a campaign
      operationId: inviteParticipant
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: participantEmail
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant invited successfully
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Forbidden - user not authorized
          content:
            text/plain:
              schema:
                type: string
                example: UserNotAuthorized
        "409":
          description: Conflict - participant already invited
          content:
            text/plain:
              schema:
                type: string
                example: Participant already invited.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/campaign/accept:
    get:
      tags:
        - RDF
      summary: Accept an invitation to a campaign
      operationId: acceptInvitation
      parameters:
        - name: participantEmail
          in: query
          required: true
          schema:
            type: string
        - name: campaignName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted successfully
          content:
            text/plain:
              schema:
                type: string
                example: Invitation accepted.
        "404":
          description: Campaign not found / participant not found
          content:
            text/plain:
              schema:
                type: string
                example: CampaignNotFound | Incorrect email.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/campaign/export:
    get:
      tags:
        - RDF
      summary: Export campaign data in turtle format
      operationId: exportCampaignData
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Campaign data exported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileName:
                    type: string
                  fileContent:
                    type: string
                    format: binary
        "401":
          description: Unauthorized - account issue
          content:
            text/plain:
              schema:
                type: string
                example: UserNotAuthorized
        "404":
          description: Campaign not found
          content:
            text/plain:
              schema:
                type: string
                example: CampaignNotFound
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Error executing script. | Oops, error. Please contact an administrator.
...